Unit tests for ExecutionService

+++ src/main/java/org/iac2/entity/compliancejob/issue/IssueFixingReportEntity.java
@@ -1,11 +1,16 @@
 package org.iac2.entity.compliancejob.issue;
 
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Null;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Data
 @Entity
@@ -13,7 +18,7 @@ import javax.validation.constraints.Null;
 public class IssueFixingReportEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+    private Long id;
 
     @NotNull
     private Boolean isSuccessful;
